Summary: Changing 'ConsoleWidget::RemoveFirstLine()' and 'conVars' type
In 'console_widget.cpp', the 'ConsoleWidget::RemoveFirstLine()' method has been changed to replace the code that removed filters that are less than two characters. Instead of using 'removeIf', a 'for' loop is now used, probably to improve compatibility with older versions of Qt.

In 'con_var.h', the type of the static variable 'conVars' has been changed from 'std::unordered_map' to 'std::map'. This change may have been made to ensure the order of the elements in the map, as 'std::map' keeps the elements in a sorted order unlike 'std::unordered_map'.